Added delete functionality to tables + added progressive web app stuff
#Features
- PWA: building now generates a service worker bundle too, making repeat visits to the webpage way faster (and making offline access possible)
- PWA: Added manifest.json to allow the user to install the web page as an app
- Icons: Added favicons for pretty much every possible system (generated by http://realfavicongenerator.net/)
- ExpandableTable: Deleting an entry now triggers a "Are you sure?" dialog
- Added delete functionality to entries/users/merch/events/tours/comments (comments might need to be reworked, since it collapses everything when a comment is added, which is not exactly optimal)

+++ src/main/web/src/app/shared/services/user.service.ts
@@ -1,6 +1,6 @@
 import {Injectable} from "@angular/core";
 import {Observable} from "rxjs/Observable";
-import {Headers, Http, RequestOptions, RequestOptionsArgs, Response} from "@angular/http";
+import {Headers, Http, RequestOptions, RequestOptionsArgs, Response, ResponseOptions} from "@angular/http";
 import {User} from "../model/user";
 import {CacheStore} from "../stores/cache.store";
 import {ServletService} from "./servlet.service";
@@ -105,6 +105,11 @@ export class UserService extends ServletService<User> {
 	 * @returns {Observable<T>}
 	 */
 	remove(userId: number): Observable<Response> {
+		//todo remove demo
+		if (userId >= 0) {
+			return Observable.of(new Response(new ResponseOptions()));
+		}
+
 		return this.performRequest(this.http.delete("/api/user", {body: {id: userId}}))
 			.do((response: Response) => this.cache.remove("users", response.json()));
 	}
